+++ frontend/index.html
@@ -11,8 +11,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Vite App</title>
   </head>
-  <body>
-    <div id="app"></div>
+  <body class="h-full">
+    <div id="app" class="flex flex-col flex-grow"></div>
     <script type="module" src="/src/main.js"></script>
   </body>
 </html>
