+++ frontend/src/assets/main.css
@@ -14,36 +14,15 @@ h6,
 p {
   font-family: "Montserrat", sans-serif;
 }
-/* #app {
-  max-width: 1280px;
-  margin: 0 auto;
-  padding: 2rem;
-  font-weight: normal;
+
+body {
+  box-sizing: border-box;
+  margin: 0;
+  padding: 0;
+  color: white;
 }
-
-a,
-.green {
-  text-decoration: none;
-  color: hsla(160, 100%, 37%, 1);
-  transition: 0.4s;
-  padding: 3px;
+#app {
+  display: flex;
+  flex-direction: column;
+  align-items: space-between;
 }
-
-@media (hover: hover) {
-  a:hover {
-    background-color: hsla(160, 100%, 37%, 0.2);
-  }
-}
-
-@media (min-width: 1024px) {
-  body {
-    display: flex;
-    place-items: center;
-  }
-
-  #app {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    padding: 0 2rem;
-  }
-} */
