+++ frontend/src/views/LandingView.vue
@@ -1,6 +1,6 @@
 <script setup>
-import { ref } from 'vue';
-const movieArr = ref(['placeholder1', 'placeholder2'])
+import { ref } from "vue";
+const movieArr = ref(["placeholder1", "placeholder2"]);
 </script>
 
 <template>
@@ -8,7 +8,7 @@ const movieArr = ref(['placeholder1', 'placeholder2'])
     <div>
       <p class="font-">Which movie do you want to watch?</p>
       <form>
-        <input type="text" placeholder="Search movie...">
+        <input type="text" placeholder="Search movie..." />
       </form>
     </div>
     <div class="flex flex-row gap-4 justify-center">
